+++ app/controllers/application_controller.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 class ApplicationController < ActionController::Base
-  include SetCurrentRequestDetails
+  include LarCity::CurrentAttributes
   include Pundit::Authorization
 
   before_action :configure_permitted_parameters, if: :devise_controller?
